rotate

+++ glap/src/cola/machine/game/myblocks/model/BoxBlock.java
@@ -150,7 +150,7 @@ public IBlock clone(){
 }
 
     @Override
-    public void render(ShaderConfig config, Vao vao, int x, int y, int z, boolean top, boolean bottom, boolean left, boolean right, boolean front, boolean back) {
+    public void render(ShaderConfig config, Vao vao, float x, float y, float z, boolean top, boolean bottom, boolean left, boolean right, boolean front, boolean back) {
 
     }
